:recycle: refactor: 약속 생성 파일 구조 변경
- Location Search 관련 파일들을 패키지로 묶어서 관리

Issue #124
